requirements.txt
altgraph==0.17.4
dlib==19.24.6
macholib==1.16.3
numpy==2.1.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pygame==2.6.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
setuptools==75.1.0
tk==0.1.0
